chore(gui): ensured read only data display keeps copying where expected - pt2
